Rename packages after structure reorganization
